Update gradle.properties
Updated mappings and API version.
API build 170, which it was set to previously, has the big issue where block loot tables are broken. I've seen multiple modders complaining of this issue in the Fabric discord, which can simply be fixed by updating the API version, so I felt it was best if the example mod was updated so people don't run into the issue in future.
I updated the mappings because I may as well.