Simply processResources logic & make it compatible with future versions of Gradle
When the processResources task is currently run, it uses deprecated Gradle features. Running "gradle clean build --warning-mode all" should output a message along the lines of "Copying or archiving duplicate paths with the default duplicates strategy has been deprecated. This is scheduled to be removed in Gradle 7.0.". This is due to inlcuding all files twice, as the "from" blocks includes them for the second time (gradle includes these files by default). This PR simply edits the relevant files from the already selected resources directory. This also reduces the amount of repetition in the script.